refactor(web): split multisig connection dropdown into separate files
